+++ test/Data/AnnotationSpec.hs
@@ -1,23 +1,27 @@
 {-# LANGUAGE DataKinds #-}
+{-# language ScopedTypeVariables #-}
 {-# LANGUAGE FlexibleContexts #-}
 {-# LANGUAGE TypeApplications #-}
+{-# LANGUAGE TypeFamilies #-}
 
 module Data.AnnotationSpec where
 
+import Data.Typeable
 import Data.Annotation
 import Test.Hspec
 
 spec :: Spec
 spec = do
-    describe "Eq" $ do
-        it "works for equal values" $ do
-            toAnnotation "hello" == toAnnotation "hello"
-        it "works for non-equal values of same type" $ do
-            toAnnotation "a" /= toAnnotation "b"
-        it "works for values of different types" $ do
-            toAnnotation (1 :: Int) /= toAnnotation "a"
     describe "Show" $ do
         it "works" $ do
             show (toAnnotation (3 :: Int))
                 `shouldBe`
                     "3"
+
+instance Eq Annotation where
+    Annotation (a :: a) == Annotation (b :: b) =
+        case eqT @a @b of
+            Just Refl ->
+                show a == show b
+            Nothing ->
+                False